Combine SEStats objects by contrast,signal

Combine SEStats objects by contrast,signal

Usage

combine_sestats(sestats_list, create_stats_df = TRUE, ...)

Arguments

sestats_list

list with two or more list objects in sestats format, as returned by se_contrast_stats().

create_stats_df

logical, default TRUE, whether to create the 'stats_df' data.frame object by merging corresponding 'stats_dfs' data.frame objects together. This step assumes that columns of class "character" or "factor" will have identical values across all data.frame objects. Specifically, it assumes that at least one column has the same name, with row identifiers sufficient to merge each result. This assumption also implies that one should not combine sestats objects when the rows are not equivalent.

...

additional arguments are ignored.

Value

list equivalent to output from se_contrast_stats() with minor exceptions:

List elements 'idesign','icontrasts','normgroup' are each returned as a list with corresponding data from each sestats object. The reason is that these values cannot logically be combined into one new object, and therefore the next level of support is to retain the underlying data.

The 'hit_array' element will contain the unique set of values for each dimension: Cutoffs, Contrasts, Signal.

Details

This function concatenates two or more objects 'sestats' as returned by se_contrast_stats(). It assumes that contrast:signal are unique across all entries, however the same contrast can be present if associated with different signal values.

The cutoffs are not considered when determining which results can be combined, in other words only the 'contrast' and 'signal' values are used to define a "unique key". Only unique contrast:signal combinations are valid as input.

In other words, input will not accept multiple sestats objects which contain the same contrast:signal and two different cutoffs.

See also