Experimental factor labels for SEDesign objects

factors() returns the labels of the underlying experimental factors, equivalent to colnames(x@design_df), i.e. the columns produced by splitting each design group name (groups(object)) on "_". factors()<- renames these labels; it has no effect on design, contrasts, groups(), or contrasts_df – it is used only for internal bookkeeping and for the print() summary.

Usage

factors(object, ...)

factors(object, ...) <- value

Arguments

object

SEDesign object

...

additional arguments are ignored

value

character vector, length equal to ncol(x@design_df) (the number of underlying experimental factors)

See also