Convert log2 fold change to signed fold change

Convert log2 fold change to signed fold change

Usage

log2fold_to_fold(x, ...)

Arguments

x

numeric vector

Value

numeric vector representing signed fold change values.

Details

This function takes log2 fold change values as input, and returns normal space fold change values that retain the positive and negative sign, and the magnitude.

For example:

log2 fold change = 2 becomes fold change = 4.

log2 fold change = -2 becomes fold change = -4.

This function therefore differs from similar functions that convert log2 fold change into a ratio. Instead, log2fold_to_fold() specifically retains the magnitude of negative changes.

See also

Examples

x <- c(-3, -2, -1, 0, 1, 2, 3);
fc <- log2fold_to_fold(x);
fc;
#> [1] -8 -4 -2  1  2  4  8

fold_to_log2fold(fc);
#> [1] -3 -2 -1  0  1  2  3