Save SE contrast stats output

Save SE contrast stats output

save_sestats(
  sestats,
  file,
  assay_names = NULL,
  contrast_names = NULL,
  cutoff_names = NULL,
  type = c("xlsx"),
  max_nchar_sheetname = 31,
  review_output = TRUE,
  sheet_prefix = NULL,
  use_assay_suffix = TRUE,
  width_factor = 1,
  max_rows = NULL,
  colorSub = NULL,
  rename_contrasts = TRUE,
  se = NULL,
  rowData_colnames = NULL,
  row_type = "gene_name",
  verbose = FALSE,
  ...
)

Arguments

sestats

list object output from se_contrast_stats()

file

character string indicating the filename to save.

assay_names

character string indicating which assay names to save, stored in dimnames(sestats$hit_array)$Signal. When NULL then all assay names are saved.

contrast_names

character string indicating which contrasts to save, stored in dimnames(sestats$hit_array)$Contrasts. When NULL then all contrasts are saved.

type

character string indicating the type of file to save.

"xlsx": saves an Excel xlsx file using jamba::writeOpenxlsx().

max_nchar_sheetname

integer number of characters allowed in MS Excel worksheet names, currently 31 characters.

review_output

logical indicating whether a summary of output should be returned as a data.frame without exporting data. This summary will indicate all worksheets to be saved, in addition to the sheetName for each worksheet.

sheet_prefix

character string with optional character prefix to use when creating worksheet names.

use_assay_suffix

logical indicating whether to include assay_names as suffix when forming sheet names, when there is more than one unique assay name to be saved. This step will attempt to abbreviate assay_names by taking up to 4 characters from each word in the assay name, where each word is delimited by "[-.:_ ]+". Otherwise, sheet names are forced to be unique by taking a substring of the contrast name of up to max_nchar_sheetname, passing any duplicate strings to jamba::makeNames() with suffix "_v" followed by an integer number.

width_factor

numeric used to adjust relative column widths in the output Excel worksheets.

colorSub

character vector of colors, optional, used to define categorical background colors for text string fields in Excel. The names(colorSub) are matched to character strings to assign colors.

verbose

logical indicating whether to print verbose output.

...

additional arguments are passed to jamba::writeOpenxlsx()

Details

This function is intended as a convenient method to export a series of statistical tables into organized, formatted Excel worksheets.