Quick conversion of hit_array to hit_list

Quick conversion of hit_array to hit_list

Usage

hit_array_to_list(
  hit_array,
  contrast_names = NULL,
  cutoff_names = NULL,
  assay_names = NULL,
  ...
)

Arguments

hit_array

array output from se_contrast_stats(), list element "hit_array".

contrast_names

character vector of contrasts.

Value

list named by contrast_names, that contains unique statistical hits by combining entries across the cutoff_names and assay_names

for each contrast.

Details

This function is mainly useful when there are multiple dimensions unresolved in a hit_array, in which case this function will combine hits across the different cutoffs and signals contained in the hit_array of sestats output from se_contrast_stats().

See also