Process sestats input into a hit incidence matrix

Process sestats input into a hit incidence matrix

Usage

process_sestats_to_hitim(
  sestats,
  cutoff_names = NULL,
  contrast_names = NULL,
  assay_names = NULL,
  contrast_suffix = NULL,
  rename_contrasts = FALSE,
  rows = NULL,
  verbose = FALSE,
  ...
)

Arguments

sestats

one of the following:

list object output from se_contrast_stats(), containing "hit_array"

array in format "hit_array" with dimnames "Cutoffs","Contrasts","Signals".

list of character vectors representing rownames(se) for the parent heatmap_se() function.

list of numeric vectors named by rownames(se).

cutoff_names, contrast_names, assay_names

character or numeric passed to hit_array_to_list() when the input is sestats or hit_array.

contrast_suffix

character optional suffix appended to the end of each contrast name.

rename_contrasts

logical indicating whether to rename contrasts by calling contrast2comp()

rows

character or NULL with optional fixed set of rownames expected in the output matrix. When rows=NULL all rows are returned using data from sestats. Otherwise, only rows defined by rows are returned.

verbose

logical indicating whether to print verbose output.

...

additional arguments are passed to contrast2comp() if relevant.

See also