Limma-voom customized for Jam

Limma-voom customized for Jam

voom_jam(
  counts,
  design = NULL,
  lib.size = NULL,
  normalize.method = "none",
  block = NULL,
  correlation = NULL,
  weights = NULL,
  span = 0.5,
  plot = FALSE,
  save.plot = TRUE,
  verbose = FALSE,
  ...
)

Details

This function is based directly upon limma::voom() with a small adjustment to handle the presence of NA values, which otherwise causes the stats::lowess() output to be clearly incorrect. The correction removes NA values during this step, producing a result as expected.

See also